requirements.txt
django==5.0
Pillow==10.4.0
mysqlclient==2.2.4
asgiref==3.8.1
sqlparse==0.5.1
djangorestframework==3.15.2
django-cors-headers==3.7.0